requirements.txt
websockets~=10.3
websocket-client~=1.4.0
pydantic~=1.10.0